requirements.txt
Django==2.0.2
fuzzywuzzy==0.16.0
djangorestframework==3.8.2
gunicorn
django-storages
psycopg2
django-heroku
dj-database-url
boto3
sendgrid
